requirements.txt
# Основные зависимости
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0

# Flask веб-приложение
flask>=3.0.0
flask-sqlalchemy>=3.1.0

# Десктопное приложение
pywebview>=4.4.0

# Для разработки (опционально)
black>=23.0.0
pylint>=3.0.0
pytest>=7.4.0

# Для визуализации (опционально)
matplotlib>=3.7.0
seaborn>=0.12.0
